expose flash methods in Page.js
